+++ apps/editor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # editor
 
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies
+  - prosemirror-suggestcat-plugin@2.2.0
+  - prosemirror-suggestcat-plugin-react@2.2.0
+
 ## 0.1.28
 
 ### Patch Changes